requirements.txt
cffi==1.14.5
cryptography==3.4.7
pycparser==2.20
pyOpenSSL==20.0.1
six==1.15.0
